fix warning

+++ RCTVideo.m
@@ -194,7 +194,7 @@ static NSString *const statusKeyPath = @"status";
     } else if(_playerItem.status == AVPlayerItemStatusFailed) {
       [_eventDispatcher sendInputEventWithName:RNVideoEventLoadingError body:@{
         @"error": @{
-          @"code": [NSNumber numberWithInt:_playerItem.error.code],
+          @"code": [NSNumber numberWithInt: (int)_playerItem.error.code],
           @"domain": _playerItem.error.domain
         },
         @"target": self.reactTag
